Temporary finalize #624 with basic tests for AixLib and MSL
