+++ bim2sim/tasks/hvac/__init__.py
@@ -6,3 +6,4 @@ from .make_graph import MakeGraph
 from .reduce import Reduce
 from .export import Export
 from .connect_elements import ConnectElements
+from .load_standardlibrary import LoadLibrariesStandardLibrary
